# apps/api/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/wallet/balances": {
      "get": {
        "operationId": "WalletReadController_balances",
        "parameters": [
          {
            "name": "ownerType",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "ownerId",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": ["Wallet"]
      }
    },
...
    "/admin/kyc/drivers": {
      "get": {
        "operationId": "AdminKycController_listDrivers",
        "parameters": [
          {
            "name": "status",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": ["Admin / KYC"]
      }
    },
    "/admin/kyc/merchants": {
      "get": {
        "operationId": "AdminKycController_listMerchants",
        "parameters": [
          {
            "name": "status",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": ["Admin / KYC"]
      }
    },
    "/admin/kyc/drivers/{id}/approve": {
      "post": {
        "operationId": "AdminKycController_approveDriver",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReviewDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": ["Admin / KYC"]
      }
    },
    "/admin/kyc/drivers/{id}/reject": {
      "post": {
        "operationId": "AdminKycController_rejectDriver",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReviewDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": ["Admin / KYC"]
      }
    },
    "/admin/kyc/merchants/{id}/approve": {
      "post": {
        "operationId": "AdminKycController_approveMerchant",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReviewDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": ["Admin / KYC"]
      }
    },
    "/admin/kyc/merchants/{id}/reject": {
      "post": {
        "operationId": "AdminKycController_rejectMerchant",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReviewDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": ["Admin / KYC"]
      }
    },
...
  },
  "info": {
    "title": "HEFA API",
    "description": "Backend endpoints for ADMIN/DRIVER/MERCHANT/USER",
    "version": "0.1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "ReviewDto": {
        "type": "object",
        "properties": {}
      },
...
    }
  }
}